[newjersey/doula-pm#84] Refactor test fields into separate files
Both cypress and unit tests want to be able to fill the entire form and progress. Previously, specified form fields for unit tests within the unit test file.

This commit refactors the specification of these test form fields into a separate file, to enable cypress tests in the future to import and re-use them.

This commit resolves #[84](https://github.com/newjersey/doula-pm/issues/84) and #[255](https://github.com/newjersey/doula-pm/issues/255).
