+++ src/app/form/(formSteps)/business-details/1/BusinessDetailsStep1.test.tsx
@@ -1,91 +1,22 @@
 import BusinessDetailsStep1 from "@/app/form/(formSteps)/business-details/1/BusinessDetailsStep1";
+import {
+  billingBusinessAddressSameAsOtherAddress,
+  businessAddressFields,
+  differentBusinessAddressSameAsOtherAddress,
+  mailingBusinessAddressSameAsOtherAddress,
+  maximalTestFields,
+  minimalTestFields,
+} from "@/app/form/(formSteps)/business-details/1/testFields";
 import type { DataStore } from "@/app/form/_utils/dataStore";
 import { renderWithProviders } from "@/app/form/_utils/testUtils/renderWithProviders";
 import {
-  createTestField,
-  createTestFields,
   testConditionalRender,
   testFillFromDataStore,
   testRequiredField,
   testSaveFieldsToDataStore,
-  type TestField,
 } from "@/app/form/_utils/testUtils/sharedTests";
 import { screen, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-
-const businessAddressQuestion = "What is your business address?";
-
-const mailingBusinessAddressSameAsOtherAddress: TestField = createTestField({
-  name: /Mailing address/i,
-  dataStoreKey: "businessAddressSameAsOtherAddress",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "radio",
-  testValue: "mailing",
-});
-const billingBusinessAddressSameAsOtherAddress: TestField = createTestField({
-  name: /Billing address/i,
-  dataStoreKey: "businessAddressSameAsOtherAddress",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "radio",
-  testValue: "billing",
-});
-const differentBusinessAddressSameAsOtherAddress: TestField = createTestField({
-  name: "I wish to enter a new address",
-  dataStoreKey: "businessAddressSameAsOtherAddress",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "radio",
-  testValue: "different",
-});
-
-const businessAddressFields = createTestFields([
-  {
-    name: "Street address *",
-    dataStoreKey: "businessStreetAddress1",
-    required: true,
-    testValue: "Test address 1",
-    withinGroupName: businessAddressQuestion,
-    prerequisiteField: differentBusinessAddressSameAsOtherAddress,
-  },
-  {
-    name: "Street address line 2",
-    dataStoreKey: "businessStreetAddress2",
-    required: false,
-    testValue: "Test address 2",
-    withinGroupName: businessAddressQuestion,
-    prerequisiteField: differentBusinessAddressSameAsOtherAddress,
-  },
-  {
-    name: "City *",
-    dataStoreKey: "businessCity",
-    required: true,
-    testValue: "Test city",
-    withinGroupName: businessAddressQuestion,
-    prerequisiteField: differentBusinessAddressSameAsOtherAddress,
-  },
-  {
-    name: "State *",
-    dataStoreKey: "businessState",
-    required: false,
-    role: "combobox",
-    testValue: "PA",
-    withinGroupName: businessAddressQuestion,
-    prerequisiteField: differentBusinessAddressSameAsOtherAddress,
-  },
-  {
-    name: "ZIP code *",
-    dataStoreKey: "businessZip",
-    required: true,
-    testValue: "12345",
-    withinGroupName: businessAddressQuestion,
-    prerequisiteField: differentBusinessAddressSameAsOtherAddress,
-  },
-]);
-
-const minimalTestFields = [mailingBusinessAddressSameAsOtherAddress];
-const allTestFields = [differentBusinessAddressSameAsOtherAddress, ...businessAddressFields];
 
 const mailingAddress = {
   streetAddress1: "123 Main St",
@@ -178,7 +109,7 @@ describe("<BusinessDetailsStep1 />", () => {
         const renderFunction = getRenderWithExistingData(mailingAddress);
         await testSaveFieldsToDataStore(
           [differentBusinessAddressSameAsOtherAddress, ...businessAddressFields],
-          allTestFields,
+          maximalTestFields,
           renderFunction,
           screen,
         );
@@ -200,7 +131,7 @@ describe("<BusinessDetailsStep1 />", () => {
         "when business address is different and $dataStoreKey is not filled in",
         async (field) => {
           const renderFunction = getRenderWithExistingData(mailingAddress);
-          await testRequiredField(field, allTestFields, renderFunction, screen);
+          await testRequiredField(field, maximalTestFields, renderFunction, screen);
         },
       );
     });
