+++ src/app/form/(formSteps)/insurance/1/InsuranceStep1.test.tsx
@@ -1,10 +1,19 @@
 import InsuranceStep1 from "@/app/form/(formSteps)/insurance/1/InsuranceStep1";
+import {
+  amountPerAggregateField,
+  amountPerOccurrenceField,
+  coverageAmountFields,
+  insuranceCoverageFields,
+  insuranceEndDateDayField,
+  insuranceEndDateYearField,
+  insuranceStartDateDayField,
+  insuranceStartDateYearField,
+  testFields,
+} from "@/app/form/(formSteps)/insurance/1/testFields";
 import type { DataStore } from "@/app/form/_utils/dataStore";
 import { getInputField } from "@/app/form/_utils/testUtils/fillInputs";
 import { renderWithProviders } from "@/app/form/_utils/testUtils/renderWithProviders";
 import {
-  createTestField,
-  createTestFields,
   type TestField,
   testFillFromDataStore,
   testInvalidField,
@@ -13,94 +22,6 @@ import {
 } from "@/app/form/_utils/testUtils/sharedTests";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-
-const insuranceStartDateDayField = createTestField({
-  name: "Day *",
-  dataStoreKey: "insuranceStartDateDay",
-  required: true,
-  testValue: "6",
-  withinGroupName: "Start date *",
-});
-
-const insuranceStartDateMonthField = createTestField({
-  name: "Month *",
-  dataStoreKey: "insuranceStartDateMonth",
-  required: true,
-  testValue: "07 - July",
-  expectedValue: "7",
-  role: "combobox",
-  withinGroupName: "Start date *",
-});
-
-const insuranceStartDateYearField = createTestField({
-  name: "Year *",
-  dataStoreKey: "insuranceStartDateYear",
-  required: true,
-  testValue: "1988",
-  withinGroupName: "Start date *",
-});
-
-const insuranceEndDateDayField = createTestField({
-  name: "Day *",
-  dataStoreKey: "insuranceEndDateDay",
-  required: true,
-  testValue: "30",
-  withinGroupName: "End date *",
-});
-
-const insuranceEndDateMonthField = createTestField({
-  name: "Month *",
-  dataStoreKey: "insuranceEndDateMonth",
-  required: true,
-  testValue: "02 - February",
-  expectedValue: "2",
-  role: "combobox",
-  withinGroupName: "End date *",
-});
-
-const insuranceEndDateYearField = createTestField({
-  name: "Year *",
-  dataStoreKey: "insuranceEndDateYear",
-  required: true,
-  testValue: "2025",
-  withinGroupName: "End date *",
-});
-
-const insuranceCoverageFields: Array<TestField> = createTestFields([
-  insuranceStartDateDayField,
-  insuranceStartDateMonthField,
-  insuranceStartDateYearField,
-  insuranceEndDateDayField,
-  insuranceEndDateMonthField,
-  insuranceEndDateYearField,
-]);
-
-const amountPerOccurrenceField: TestField = {
-  name: "Amount per occurrence *",
-  dataStoreKey: "insuranceOccurenceAmount",
-  requiredErrorMessage: "Amount per occurrence is required",
-  role: "textbox",
-  required: true,
-  testValue: "1000005",
-  expectedValue: "1000005",
-};
-
-const amountPerAggregateField: TestField = {
-  name: "Amount per aggregate *",
-  dataStoreKey: "insuranceAggregateAmount",
-  requiredErrorMessage: "Amount per aggregate is required",
-  role: "textbox",
-  required: true,
-  testValue: "3000300",
-  expectedValue: "3000300",
-};
-
-const coverageAmountFields: Array<TestField> = createTestFields([
-  amountPerOccurrenceField,
-  amountPerAggregateField,
-]);
-
-const testFields: Array<TestField> = [...insuranceCoverageFields, ...coverageAmountFields];
 
 describe("<InsuranceStep1 />", () => {
   const renderFunction = (dataStore: DataStore = {}) =>
