+++ src/app/form/(formSteps)/personal-details/2/PersonalDetailsStep2.test.tsx
@@ -1,11 +1,19 @@
 import PersonalDetailsStep2 from "@/app/form/(formSteps)/personal-details/2/PersonalDetailsStep2";
+import {
+  billingAddressFields,
+  mailingAddressFields,
+  mailingAddressQuestion,
+  minimalTestFields,
+  noSameBillingMailingAddress,
+  testFields,
+  yesSameBillingMailingAddress,
+  zipCodeField,
+} from "@/app/form/(formSteps)/personal-details/2/testFields";
 import type { DataStore } from "@/app/form/_utils/dataStore";
 import { expectAddressHasAutocomplete } from "@/app/form/_utils/testUtils/autocomplete";
 import { getInputField } from "@/app/form/_utils/testUtils/fillInputs";
 import { renderWithProviders } from "@/app/form/_utils/testUtils/renderWithProviders";
 import {
-  createTestField,
-  createTestFields,
   testConditionalRender,
   type TestField,
   testFillFromDataStore,
@@ -15,129 +23,6 @@ import {
 } from "@/app/form/_utils/testUtils/sharedTests";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-
-const mailingAddressQuestion =
-  "Mailing address We will send official mail here. It can be your home address.";
-const billingAddressQuestion = "What is your billing address?";
-
-const mailingAddressFields = createTestFields([
-  {
-    name: "Street address *",
-    dataStoreKey: "streetAddress1",
-    required: true,
-    testValue: "Test address 1",
-    withinGroupName: mailingAddressQuestion,
-  },
-  {
-    name: "Street address line 2",
-    dataStoreKey: "streetAddress2",
-    required: false,
-    testValue: "Test address 2",
-    withinGroupName: mailingAddressQuestion,
-  },
-  {
-    name: "City *",
-    dataStoreKey: "city",
-    required: true,
-    testValue: "Test city",
-    withinGroupName: mailingAddressQuestion,
-  },
-  {
-    name: "State *",
-    dataStoreKey: "state",
-    required: false,
-    role: "combobox",
-    testValue: "PA",
-    withinGroupName: mailingAddressQuestion,
-  },
-  {
-    name: "ZIP code *",
-    dataStoreKey: "zip",
-    required: true,
-    testValue: "12345",
-    withinGroupName: mailingAddressQuestion,
-  },
-]);
-
-const yesSameBillingMailingAddress: TestField = {
-  name: "Yes",
-  dataStoreKey: "hasSameBillingMailingAddress",
-  required: true,
-  requiredErrorMessage: "This question is required",
-  role: "radio",
-  testValue: "true",
-  expectedValue: "true",
-  withinGroupName: "Are your billing and residential addresses the same? Select one *",
-};
-const noSameBillingMailingAddress: TestField = {
-  name: "No",
-  dataStoreKey: "hasSameBillingMailingAddress",
-  required: true,
-  requiredErrorMessage: "This question is required",
-  role: "radio",
-  testValue: "false",
-  expectedValue: "false",
-  withinGroupName: "Are your billing and residential addresses the same? Select one *",
-};
-
-const minimalTestFields = [...mailingAddressFields, yesSameBillingMailingAddress];
-
-const zipCodeField = createTestField({
-  name: "ZIP code *",
-  dataStoreKey: "billingZip",
-  required: true,
-  testValue: "12345",
-  withinGroupName: billingAddressQuestion,
-  alternateRequiredFieldError: "Billing zip code is required",
-  prerequisiteField: noSameBillingMailingAddress,
-});
-
-const billingAddressFields = [
-  ...createTestFields([
-    {
-      name: "Street address *",
-      dataStoreKey: "billingStreetAddress1",
-      required: true,
-      testValue: "Test address 1",
-      withinGroupName: billingAddressQuestion,
-      alternateRequiredFieldError: "Billing street address is required",
-      prerequisiteField: noSameBillingMailingAddress,
-    },
-    {
-      name: "Street address line 2",
-      dataStoreKey: "billingStreetAddress2",
-      required: false,
-      testValue: "Test address 2",
-      withinGroupName: billingAddressQuestion,
-      prerequisiteField: noSameBillingMailingAddress,
-    },
-    {
-      name: "City *",
-      dataStoreKey: "billingCity",
-      required: true,
-      testValue: "Houston",
-      withinGroupName: billingAddressQuestion,
-      alternateRequiredFieldError: "Billing city is required",
-      prerequisiteField: noSameBillingMailingAddress,
-    },
-    {
-      name: "State *",
-      dataStoreKey: "billingState",
-      required: false,
-      role: "combobox",
-      testValue: "TX",
-      withinGroupName: billingAddressQuestion,
-      prerequisiteField: noSameBillingMailingAddress,
-    },
-  ]),
-  zipCodeField,
-];
-
-const allTestFields = [
-  ...mailingAddressFields,
-  noSameBillingMailingAddress,
-  ...billingAddressFields,
-];
 
 describe("<PersonalDetailsStep2 />", () => {
   const renderFunction = (dataStore: DataStore = {}) =>
@@ -188,7 +73,7 @@ describe("<PersonalDetailsStep2 />", () => {
       await testInvalidField(
         { ...zipCodeField, testValue: "1" },
         "Billing zip code must have five digits",
-        allTestFields,
+        testFields,
         renderFunction,
         screen,
       );
@@ -221,7 +106,7 @@ describe("<PersonalDetailsStep2 />", () => {
       it("when billing address is different from mailing address", async () => {
         await testSaveFieldsToDataStore(
           [noSameBillingMailingAddress, ...billingAddressFields],
-          allTestFields,
+          testFields,
           renderFunction,
           screen,
         );
@@ -241,7 +126,7 @@ describe("<PersonalDetailsStep2 />", () => {
       it.each(billingAddressFields.filter((field) => field.required))(
         "when mailing and billing address are different and $dataStoreKey is not filled in",
         async (field: TestField) => {
-          await testRequiredField(field, allTestFields, renderFunction, screen);
+          await testRequiredField(field, testFields, renderFunction, screen);
         },
       );
     });
