+++ src/app/form/(formSteps)/screening/3/ScreeningStep3.test.tsx
@@ -1,60 +1,23 @@
 import ScreeningStep3 from "@/app/form/(formSteps)/screening/3/ScreeningStep3";
+import {
+  testFields,
+  yesHadDhmasBusiness,
+  yesHaveOtherBusinessOwnerNextYear,
+} from "@/app/form/(formSteps)/screening/3/testFields";
 import type { DataStore } from "@/app/form/_utils/dataStore";
 import { renderWithProviders } from "@/app/form/_utils/testUtils/renderWithProviders";
 import {
-  createTestField,
   testInvalidField,
   testSaveFieldsToDataStore,
-  type TestField,
 } from "@/app/form/_utils/testUtils/sharedTests";
 import { screen } from "@testing-library/react";
-
-const noHaveOtherBusinessOwnerNextYear = createTestField({
-  name: "No",
-  role: "radio",
-  required: true,
-  dataStoreKey: "haveOtherBusinessOwnerNextYear",
-  testValue: "false",
-  withinGroupName:
-    "Do you anticipate anyone else having a percentage of your business in the next year? Select one *",
-});
-const yesHaveOtherBusinessOwnerNextYear = createTestField({
-  name: "Yes",
-  role: "radio",
-  required: true,
-  dataStoreKey: "haveOtherBusinessOwnerNextYear",
-  testValue: "true",
-  withinGroupName:
-    "Do you anticipate anyone else having a percentage of your business in the next year? Select one *",
-});
-
-const noHadDhmasBusiness = createTestField({
-  name: "No",
-  role: "radio",
-  required: true,
-  dataStoreKey: "hadDhmasBusiness",
-  testValue: "false",
-  withinGroupName:
-    "In the last 5 years, have you owned any percentage of companies that do business with the Division of Medical Assistance and Health Services? Select one *",
-});
-const yesHadDhmasBusiness = createTestField({
-  name: "Yes",
-  role: "radio",
-  required: true,
-  dataStoreKey: "hadDhmasBusiness",
-  testValue: "true",
-  withinGroupName:
-    "In the last 5 years, have you owned any percentage of companies that do business with the Division of Medical Assistance and Health Services? Select one *",
-});
-
-const allTestFields: Array<TestField> = [noHaveOtherBusinessOwnerNextYear, noHadDhmasBusiness];
 
 describe("<ScreeningStep3 />", () => {
   const renderFunction = (dataStore: DataStore = {}) =>
     renderWithProviders(<ScreeningStep3 />, "/form/screening/3", dataStore);
 
   it("saves fields to the data store on submit", async () => {
-    await testSaveFieldsToDataStore(allTestFields, allTestFields, renderFunction, screen);
+    await testSaveFieldsToDataStore(testFields, testFields, renderFunction, screen);
   });
 
   it.each([[yesHaveOtherBusinessOwnerNextYear], [yesHadDhmasBusiness]])(
@@ -63,7 +26,7 @@ describe("<ScreeningStep3 />", () => {
       await testInvalidField(
         invalidField,
         "Currently this site cannot support your situation. Please use the standard FFS application",
-        allTestFields,
+        testFields,
         renderFunction,
         screen,
         invalidField,
