+++ src/app/form/(formSteps)/training/1/TrainingStep1.test.tsx
@@ -1,9 +1,18 @@
+import {
+  childrensFuturesTrainingOrganization,
+  maximalTestFields,
+  minimalTestFields,
+  nameOfTrainingOrganization,
+  noDoulaTrainingInPerson,
+  noneTrainingOrganization,
+  trainingAddressFields,
+  trainingInstructorFields,
+  yesDoulaTrainingInPerson,
+} from "@/app/form/(formSteps)/training/1/testFields";
 import TrainingStep1 from "@/app/form/(formSteps)/training/1/TrainingStep1";
 import type { DataStore } from "@/app/form/_utils/dataStore";
 import { renderWithProviders } from "@/app/form/_utils/testUtils/renderWithProviders";
 import {
-  createTestField,
-  createTestFields,
   testConditionalRender,
   testFillFromDataStore,
   testRequiredField,
@@ -13,145 +22,6 @@ import {
 import { fillField } from "@form/_utils/testUtils/fillInputs";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-
-const trainingAddressGroupName = "What is the address of your training organization? *";
-
-const childrensFuturesTrainingOrganization: TestField = createTestField({
-  name: "Which state-approved training did you complete? Select one *",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "combobox",
-  testValue: "Children's Futures (Trenton)",
-  dataStoreKey: "stateApprovedTraining",
-});
-
-const noneTrainingOrganization: TestField = createTestField({
-  name: "Which state-approved training did you complete? Select one *",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "combobox",
-  testValue: "None of these",
-  dataStoreKey: "stateApprovedTraining",
-});
-
-const nameOfTrainingOrganization: TestField = createTestField({
-  name: "What is the name of your training organization? *",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "textbox",
-  testValue: "Test organization",
-  dataStoreKey: "nameOfTrainingOrganization",
-  prerequisiteField: noneTrainingOrganization,
-});
-
-const yesDoulaTrainingInPerson: TestField = createTestField({
-  name: "Yes, in person or hybrid",
-  dataStoreKey: "isDoulaTrainingInPerson",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "radio",
-  testValue: "true",
-  withinGroupName: "Did you attend your doula training classes in person? Select one *",
-});
-
-const noDoulaTrainingInPerson: TestField = createTestField({
-  name: "No, it was virtual",
-  dataStoreKey: "isDoulaTrainingInPerson",
-  required: true,
-  alternateRequiredFieldError: "This question is required",
-  role: "radio",
-  testValue: "false",
-  withinGroupName: "Did you attend your doula training classes in person? Select one *",
-});
-
-const trainingAddressFields: TestField[] = createTestFields([
-  {
-    name: "Street address *",
-    required: true,
-    dataStoreKey: "trainingStreetAddress1",
-    alternateRequiredFieldError: "Training street address is required",
-    testValue: "Test address 1",
-    withinGroupName: trainingAddressGroupName,
-    prerequisiteField: yesDoulaTrainingInPerson,
-  },
-  {
-    name: "Street address line 2",
-    required: false,
-    dataStoreKey: "trainingStreetAddress2",
-    testValue: "Test address 2",
-    withinGroupName: trainingAddressGroupName,
-    prerequisiteField: yesDoulaTrainingInPerson,
-  },
-  {
-    name: "City *",
-    required: true,
-    alternateRequiredFieldError: "Training city is required",
-    dataStoreKey: "trainingCity",
-    testValue: "Test city",
-    withinGroupName: trainingAddressGroupName,
-    prerequisiteField: yesDoulaTrainingInPerson,
-  },
-  {
-    name: "State *",
-    required: false,
-    alternateRequiredFieldError: "Training state is required",
-    role: "combobox",
-    testValue: "NJ",
-    dataStoreKey: "trainingState",
-    withinGroupName: trainingAddressGroupName,
-    prerequisiteField: yesDoulaTrainingInPerson,
-  },
-  {
-    name: "ZIP code *",
-    required: true,
-    alternateRequiredFieldError: "Training zip code is required",
-    dataStoreKey: "trainingZip",
-    testValue: "12345",
-    withinGroupName: trainingAddressGroupName,
-    prerequisiteField: yesDoulaTrainingInPerson,
-  },
-]);
-
-const trainingInstructorFields: TestField[] = createTestFields([
-  {
-    name: "First name *",
-    required: true,
-    dataStoreKey: "instructorFirstName",
-    testValue: "Jane",
-  },
-  {
-    name: "Last name *",
-    required: true,
-    dataStoreKey: "instructorLastName",
-    testValue: "Doe",
-  },
-  {
-    name: "Email address *",
-    required: true,
-    dataStoreKey: "instructorEmail",
-    testValue: "test@example.com",
-  },
-  {
-    name: "Phone number",
-    required: false,
-    dataStoreKey: "instructorPhoneNumber",
-    testValue: "111-111-1111",
-  },
-]);
-
-const minimalTestFields = [
-  childrensFuturesTrainingOrganization,
-  noDoulaTrainingInPerson,
-  ...trainingInstructorFields,
-];
-
-const allTestFields = [
-  noneTrainingOrganization,
-  nameOfTrainingOrganization,
-  yesDoulaTrainingInPerson,
-  ...trainingAddressFields,
-  ...trainingInstructorFields,
-];
 
 const renderFunction = (dataStore: DataStore = {}) =>
   renderWithProviders(<TrainingStep1 />, "/form/training/1", dataStore);
@@ -171,7 +41,7 @@ describe("<TrainingStep1 />", () => {
       it("when 'None of these' is selected and the training organization name is provided", async () => {
         await testSaveFieldsToDataStore(
           [noneTrainingOrganization, nameOfTrainingOrganization],
-          allTestFields,
+          maximalTestFields,
           renderFunction,
           screen,
         );
@@ -188,7 +58,12 @@ describe("<TrainingStep1 />", () => {
         );
       });
       it("when None of the these is selected and nameOfTrainingOrganization is not filled in", async () => {
-        await testRequiredField(nameOfTrainingOrganization, allTestFields, renderFunction, screen);
+        await testRequiredField(
+          nameOfTrainingOrganization,
+          maximalTestFields,
+          renderFunction,
+          screen,
+        );
       });
     });
 
@@ -236,7 +111,7 @@ describe("<TrainingStep1 />", () => {
       it("when training was in person or hybrid", async () => {
         await testSaveFieldsToDataStore(
           [yesDoulaTrainingInPerson, ...trainingAddressFields],
-          allTestFields,
+          maximalTestFields,
           renderFunction,
           screen,
         );
@@ -255,7 +130,7 @@ describe("<TrainingStep1 />", () => {
       it.each(trainingAddressFields.filter((field) => field.required))(
         "when training was in person or hybrid and $dataStoreKey is not filled in",
         async (field: TestField) => {
-          await testRequiredField(field, allTestFields, renderFunction, screen);
+          await testRequiredField(field, maximalTestFields, renderFunction, screen);
         },
       );
     });
@@ -263,7 +138,7 @@ describe("<TrainingStep1 />", () => {
     it.each(trainingAddressFields.filter((field) => field.required))(
       "marks $dataStoreKey as required and displays an error message if it is not filled in",
       async (field: TestField) => {
-        await testRequiredField(field, allTestFields, renderFunction, screen);
+        await testRequiredField(field, maximalTestFields, renderFunction, screen);
       },
     );
